+++ contracts/interfaces/IDebtToken.sol
@@ -4,6 +4,7 @@ pragma solidity ^0.8.0;
 import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IDebtToken is IERC20 {
-    function mint(address _to, uint _amount) external;
-    function burn(uint _amount) external;
+    function mint(address _to, uint256 _amount) external;
+
+    function burn(uint256 _amount) external;
 }
