+++ contracts/interfaces/IOracle.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.8.0;
 
 interface IOracle {
-    function fetchPrice() external view returns (bool, uint);
+    function fetchPrice() external view returns (bool, uint256);
 }
